mask signals while using connect(2) in connectto()
This makes sure connect(2) don't return a possible < 0 and errno = EINTR.

POSIX says for connect(2): "
[EINTR]         The connection attempt was interrupted by a signal.
                The attempt will continue asynchronously as if the
                socket was non-blocking."

This was reproducable by launching with st -e sacc bitreich.org . When sacc was
signaled with SIGWINCH while it was connecting the connect(2) would be handled
as a failure and the next interface was tried. This gave a vague message "no
route to host", because the next interfaces could fail.